requirements.txt
annotated-types==0.7.0
anyio==4.5.0
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.1
exceptiongroup==1.2.2
fastapi==0.115.0
h11==0.14.0
idna==3.10
iniconfig==2.0.0
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pytest==8.3.3
requests==2.32.3
sniffio==1.3.1
starlette==0.38.6
tomli==2.0.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
